final_pjt_back/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
